feat(passwordReset): password request&password reset done.

+++ client/app/password/password.controller.js
@@ -3,10 +3,8 @@
 var app = angular.module('jandiApp');
 
 app.controller('passwordResetController', function($scope, $state, $filter, $location, $modal, authAPIservice, $rootScope, $timeout, storageAPIservice) {
-    var teamId = $state.params.teamId;
     var token = $state.params.token;
 
-    var userId = '';
     $scope.token = '';
 
     $scope.title = $filter('translate')('@common-new-password');
@@ -14,25 +12,29 @@ app.controller('passwordResetController', function($scope, $state, $filter, $loc
     $scope.resetDone = false;
     $scope.resetFailed = false;
 
-    authAPIservice.validatePasswordToken(teamId, token)
-        .success(function(response) {
-            userId = response.id;
-            token = response.token;
-        })
-        .error(function(response) {
-            //alert($filter('translate')('@common-invalid-request'));
-            //$state.go('signin');
-        });
+    //authAPIservice.validatePasswordToken(teamId, token)
+    //    .success(function(response) {
+    //        userId = response.id;
+    //        token = response.token;
+    //    })
+    //    .error(function(response) {
+    //        //alert($filter('translate')('@common-invalid-request'));
+    //        //$state.go('signin');
+    //    });
+
 
     $scope.resetPassword = function(user) {
+
+        if ($scope.isLoading) return;
+
         $scope.isLoading = true;
-        authAPIservice.resetPassword(teamId, token, user.password, $rootScope.preferences.serverLang)
+
+        authAPIservice.resetPassword(token, user.password)
             .success(function(response) {
 
                 $scope.resetDone = true;
-                $scope.isLoading = false;
 
-                storageAPIservice.removeAccessToken();
+                storageAPIservice.removeLocal();
                 storageAPIservice.removeSession();
 
                 $timeout(function() {
@@ -40,8 +42,16 @@ app.controller('passwordResetController', function($scope, $state, $filter, $loc
                 }, 3000)
             })
             .error(function(response) {
-                $scope.isLoading = false;
                 $scope.resetFailed = true;
+                if (response.data.code == 40000) {
+                    // Invalid token
+                    alert($filter('translate')('@common-invalid-request'));
+                }
+
+            })
+            .finally(function() {
+                $scope.isLoading = false;
+
             })
     };
 });
